+++ class/dolsendinblue.class.php
@@ -381,7 +381,7 @@ class DolSendinBlue extends CommonObject
 		dol_syslog(get_class($this) . "::fetch_by_sendinblueid sql=" . $sql, LOG_DEBUG);
 		$resql = $this->db->query($sql);
 		if ($resql) {
-			if ($this->db->num_rows($resql)) {
+			if ($this->db->num_rows($resql) > 0) {
 				$obj = $this->db->fetch_object($resql);
 
 				$this->id = $obj->rowid;
@@ -397,10 +397,13 @@ class DolSendinBlue extends CommonObject
 				$this->datec = $this->db->jdate($obj->datec);
 				$this->fk_user_mod = $obj->fk_user_mod;
 				$this->tms = $this->db->jdate($obj->tms);
-			}
-			$this->db->free($resql);
 
-			return 1;
+
+				$this->db->free($resql);
+				return 1;
+			}
+
+			return 0;
 		} else {
 			$this->error = "Error " . $this->db->lasterror();
 			dol_syslog(get_class($this) . "::fetch_by_sendinblueid " . $this->error, LOG_ERR);
@@ -1404,10 +1407,48 @@ class DolSendinBlue extends CommonObject
 			dol_syslog(get_class($this) . '::addEmailToList count($batch_email_to_add)=' . count($batch_email_to_add), LOG_DEBUG);
 
 			dol_syslog(get_class($this) . '::addEmailToList start batchSubscribe ' . dol_print_date(dol_now(), 'standard'), LOG_DEBUG);
+//
+//			foreach($batch_email_to_add as $email){
+//				// Call
+//				$data = array(
+//					"email" => $email['email_address'],
+//					"attributes" => array(
+//						"FIRST_NAME" => $email['merge_vars']->FNAME,
+//						"LAST_NAME" => $email['merge_vars']->LNAME,
+//					),
+//					"listIds" => array(intval($listid))
+//				);
+//
+//				if (!empty($email['merge_extrafields']))
+//				{
+//					foreach ($email['merge_extrafields'] as $code => $val)
+//					{
+//						// strtoupper car sendinblue force les majuscules et remplace les espaces par des _
+//						$data['attributes'][strtoupper($code)] = $val;
+//					}
+//				}
+//
+//				try {
+//					$response = $this->sendinblue->create_update_user($data);
+//					if($this->sendinblue->analyseResponseResult($response)){
+//						$this->errors[] = $this->sendinblue->error;
+//						$error ++;
+//					}
+//				} catch ( Exception $e ) {
+//					$this->errors[] = $e->getMessage();
+//					$batch_email_to_add_error=$batch_email_to_add;
+//					$error ++;
+//				}
+//
+//			}
+
+			$contactToAddInList = array();
+			$contactData = array();
 
 			foreach($batch_email_to_add as $email){
-				// Call
+				$contactToAddInList[] = $email['email_address'];
 
+				// Call
 				$data = array(
 					"email" => $email['email_address'],
 					"attributes" => array(
@@ -1426,15 +1467,31 @@ class DolSendinBlue extends CommonObject
 					}
 				}
 
-				try {
-					$response = $this->sendinblue->create_update_user($data);
-				} catch ( Exception $e ) {
-					$this->errors[] = $e->getMessage();
-					$batch_email_to_add_error=$batch_email_to_add;
+				// If update fail we will create contact with this info
+				$contactData[$email['email_address']] = $data;
+			}
+
+
+			try {
+
+				$response = $this->sendinblue->addExistingContactsToLists($listid, array('emails' => $contactToAddInList));
+				if($this->sendinblue->analyseResponseResult($response)){
+					$this->errors[] = $this->sendinblue->error;
 					$error ++;
 				}
-
+				elseif(!empty($response['failure']) && is_array($response['failure'])){
+					foreach ($response['failure'] as $email ){
+						if(isset($contactData[$email])){
+							$response = $this->sendinblue->create_update_user($contactData[$email]);
+						}
+					}
+				}
+			} catch ( Exception $e ) {
+				$this->errors[] = $e->getMessage();
+				$batch_email_to_add_error=$batch_email_to_add;
+				$error ++;
 			}
+
 
 			dol_syslog(get_class($this) . '::addEmailToList end batchSubscribe ' . dol_print_date(dol_now(), 'standard'), LOG_DEBUG);
 		}
@@ -2826,7 +2883,7 @@ class DolSendinBlue extends CommonObject
 			$this->getInstanceSendinBlue();
 			$reponse = $this->sendinblue->get_user($email);
 
-			// TODO : change hard_bounces does not exist anymore...
+			// TODO : change hard_bounces does not exist anymore... --> use webhooks
 			if(!empty($reponse['statistics']['hardBounces'])){
 				foreach($reponse['statistics']['hardBounces'] as $camp){
 					$fk_mailing = $this->getMaillingFromCampainId($camp['campaignId']);
@@ -2840,7 +2897,7 @@ class DolSendinBlue extends CommonObject
 				}
 			}
 
-			// TODO : change soft_bounces does not exist anymore
+			// TODO : change soft_bounces does not exist anymore --> use webhooks
 			if(!empty($reponse['statistics']['softBounces'])){
 				foreach($reponse['statistics']['softBounces'] as $camp){
 					$fk_mailing = $this->getMaillingFromCampainId($camp['campaignId']);
@@ -2853,7 +2910,7 @@ class DolSendinBlue extends CommonObject
 					}
 				}
 			}
-			// TODO change spam does not exist anymore
+			// TODO change spam does not exist anymore --> use webhooks
 			if(!empty($reponse['statistics']['spam'])){
 				foreach($reponse['statistics']['spam'] as $camp){
 					$fk_mailing = $this->getMaillingFromCampainId($camp['campaignId']);